Added support for Leaflet 1.0+
Starting in Leaflet 1.0, the L.Control class stores current layers as an
Array ([]) versus an id-mapped Object ({}).  This caused issues with
the current codebase as it expected to look up layers by their ID
using a simple property lookup.

In this commit I use the new Array style for storing layers but added
a helper method `_getLayer()` that is used to retrieve a layer by its
ID.  I replaced all layer accesses with this new method, which allows
the library to maintain both 0.7.x and 1.0+ compatibility.

To promote moving forward, I updated the examples to use the latest
Leaflet version (1.0.1).  To help verify compatiblity, I moved a copy
of the existing examples to a 0.7.x/ folder.  I also updated the CDN
in the examples to use cdnjs.com (since the existing host did not
contain needed 1.0 libraries).