Collapsable Groups.
Added a config option "groupsCollapsable" which if set to true, will
make each group header individually collapsable.  This is useful when
dealing with either small map elements, or a larger number of layers
within the groups.

Also added two supporting config settings: "groupsCollapseClass" and
"groupsExpandClass" which control the css styles used for the toggle
indicators, allowing users to override the default +/- indicators.
Setting one to "glyphicon glyphicon-chevron-right" will use the
glyphicon provided with bootstrap for example.